requirements.txt
# Core Application
fastapi==0.111.0
uvicorn[standard]==0.29.0
discord.py==2.3.2
python-dotenv==1.0.1
aiohttp==3.9.5
Pillow==10.3.0
pytesseract==0.3.10
multiprocessing-logging==0.3.1

# AI & Data Processing
openai==1.30.1
pinecone-client==4.1.1
sentence-transformers==2.7.0
spacy==3.7.5
numpy==1.26.4

# Document Parsing (Install all extras for comprehensive support)
unstructured[all]==0.14.4
python-pptx==0.6.23 # Explicit dependency for .pptx
python-magic==0.4.27 # For MIME type detection

# Testing & Linting
pytest==8.2.1
pytest-asyncio==0.23.7 